+++ example/lib/main.dart
@@ -25,7 +25,7 @@ class MyApp extends StatelessWidget {
 
     // launch the FML engine
     return FmlEngine(
-            ApplicationType.single,
+            ApplicationType.multi,
             domain: example1,
             title: "Flutter Markup Language V$version",
             version: version,
