+++ lib/application/application_manager.dart
@@ -12,7 +12,7 @@ import 'package:fml/event/event.dart';
 import 'package:fml/helpers/helpers.dart';
 
 // platform
-import 'package:fml/platform/platform.web.dart'
+import 'package:fml/platform/platform.vm.dart'
     if (dart.library.io) 'package:fml/platform/platform.vm.dart'
     if (dart.library.html) 'package:fml/platform/platform.web.dart';
