+++ lib/log/manager.dart
@@ -6,7 +6,7 @@ import 'package:fml/hive/log.dart' as hive_log;
 import 'package:fml/helpers/helpers.dart';
 
 // platform
-import 'package:fml/platform/platform.web.dart'
+import 'package:fml/platform/platform.vm.dart'
     if (dart.library.io) 'package:fml/platform/platform.vm.dart'
     if (dart.library.html) 'package:fml/platform/platform.web.dart';
