+++ lib/widgets/camera/stream/stream.web.dart
@@ -21,7 +21,7 @@ import 'package:fml/datasources/detectors/image/detectable_image.stub.dart'
     if (dart.library.html) 'package:fml/datasources/detectors/image/detectable_image.web.dart';
 
 // platform
-import 'package:fml/platform/platform.web.dart'
+import 'package:fml/platform/platform.vm.dart'
     if (dart.library.io) 'package:fml/platform/platform.vm.dart'
     if (dart.library.html) 'package:fml/platform/platform.web.dart';
