+++ lib/widgets/colorpicker/colorpicker_view.dart
@@ -1,6 +1,4 @@
 // © COPYRIGHT 2022 APPDADDY SOFTWARE SOLUTIONS INC. ALL RIGHTS RESERVED.
-import 'dart:async';
-
 import 'package:flex_color_picker/flex_color_picker.dart';
 import 'package:flutter/material.dart';
 import 'package:fml/widgets/colorpicker/colorpicker_model.dart';
