+++ lib/widgets/framework/framework_view.dart
@@ -21,7 +21,7 @@ import 'package:fml/helpers/helpers.dart';
 import 'package:fml/phrase.dart';
 
 // platform
-import 'package:fml/platform/platform.web.dart'
+import 'package:fml/platform/platform.vm.dart'
     if (dart.library.io) 'package:fml/platform/platform.vm.dart'
     if (dart.library.html) 'package:fml/platform/platform.web.dart';
