+++ lib/widgets/table/table_model.dart
@@ -24,7 +24,7 @@ import 'package:fml/helpers/helpers.dart';
 import 'package:fml/helpers/mime.dart';
 
 // platform
-import 'package:fml/platform/platform.web.dart'
+import 'package:fml/platform/platform.vm.dart'
     if (dart.library.io) 'package:fml/platform/platform.vm.dart'
     if (dart.library.html) 'package:fml/platform/platform.web.dart';
